+++ packages/file-storage/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`file-storage`](https://github.com/remix-run/remix/tree/v3/packages/file-storage). It follows [semantic versioning](https://semver.org/).
 
+## HEAD
+
+- Remove hash directories when they are empty in `LocalFileStorage`
+
 ## v0.8.0 (2025-07-21)
 
 - Renamed package from `@mjackson/file-storage` to `@remix-run/file-storage`